requirements.txt
pyserial
aiohttp
PYyaml